paper modal moved to own component
